Cohesion for Admin Route and Controller
